deprecated/mnist/pytorch/requirements.serve.txt
fastapi[all]
python-multipart
pillow
vessl==0.1.143
numpy
pandas
torch==2.0.1
torchvision==0.15.2